fileupload_reqrm/requirements.txt
Flask==2.0.1
azure-storage-blob==12.9.0
Werkzeug ==2.0.0
